Create Find the Missing Number.py
Method 1: This method uses the technique of the Summation formula. 

Approach: The length of the array is n-1. So, the sum of all n elements i.e. sum of numbers from 1 to n can be calculated using the formula n*(n+1)/2. Now find the sum of all the elements in the array and subtract it from the sum of the first n natural numbers, it will give us the value of the missing element.
Algorithm: 
Calculate the sum of the first n natural numbers as sumtotal= n*(n+1)/2
Create a variable sum to store the sum of the array elements.
Traverse the array from start to end.
Update the value of sum as sum = sum + array[i]
Print the missing number as SumTotal – sum